grandpa: rename serialized to report in rpc module
